Refactor reducers’s file names under Surveys screen
Replace singulars with plurals.
