+++ src/screens/Surveys/screens/SurveyDetails/reducers.js
@@ -17,7 +17,7 @@ import {
 } from './data/survey/actions';
 
 // Reducers
-import dataReducer from './data/reducer';
+import dataReducer from './data/reducers';
 
 // Constants
 import STATE_MODELS from '../../../../constants/models/state';
