+++ src/screens/Surveys/screens/SurveyDoorway/data/reducers.js
@@ -2,7 +2,7 @@
 import { combineReducers } from 'redux';
 
 // Reducers
-import landingReducer from './landing/reducer';
+import landingReducer from './landing/reducers';
 
 // Combine reducers
 export default combineReducers({
