+++ src/screens/Surveys/screens/SurveyNew/reducers.js
